feat(exception): 에러를 담는 객체 생성
