--- server/src/main/java/com/onetool/server/global/new_exception/exception/BaseException.java
@@ -1,32 +1,0 @@
-package com.whale.gather_one.global.exception;
-
-import com.whale.gather_one.global.exception.error.ErrorCode;
-import lombok.Getter;
-
-@Getter
-public class BaseException extends RuntimeException {
-
-    private final ErrorCode errorCode;
-    private String customErrorMessage;
-
-    private BaseException(ErrorCode code) {
-        this.errorCode = code;
-    }
-
-    private BaseException(ErrorCode code, final String message) {
-        this.errorCode = code;
-        this.customErrorMessage = message;
-    }
-
-    public boolean hasCustomMessage(){
-        return customErrorMessage != null;
-    }
-
-    public static BaseException from(ErrorCode errorCode) {
-        return new BaseException(errorCode);
-    }
-
-    public static BaseException from(ErrorCode errorCode, final String customErrorMessage){
-        return new BaseException(errorCode, customErrorMessage);
-    }
-}
